backend/requirements.txt
fastapi>=0.135.3
uvicorn>=0.44.0
langchain-chroma>=0.1.2
langchain-community>=0.4.1
langchain-groq>=1.1.2
langchain-huggingface>=1.2.1
langchain-text-splitters>=0.2.0
python-dotenv>=1.0.1
pydantic>=2.0.0